выделили из шаблонов в отдельные модули header и footer, и данные для них в структуры данных
