+++ js/data/data.js
@@ -1,0 +1,8 @@
+export const footerData = {
+  creationDate: 2016
+};
+
+export const gameState = {
+  lives: 1,
+  time: 30
+};
