+++ js/main.js
@@ -1,4 +1,4 @@
 import renderScreen from './modules/utils/render-screen';
-import intro from './modules/screens/intro';
+import intro from './modules/game/intro';
 
 renderScreen(intro);
