+++ js/modules/game/game3.js
@@ -3,20 +3,11 @@ import renderScreen from '../utils/render-screen';
 import nextScreen from './stats';
 import intro from './intro';
 
-const template = String.raw`<header class="header">
-<div class="header__back">
-  <button class="back">
-    <img src="img/arrow_left.svg" width="45" height="45" alt="Back">
-    <img src="img/logo_small.svg" width="101" height="44">
-  </button>
-</div>
-<h1 class="game__timer">NN</h1>
-<div class="game__lives">
-  <img src="img/heart__empty.svg" class="game__heart" alt="Life" width="32" height="32">
-  <img src="img/heart__full.svg" class="game__heart" alt="Life" width="32" height="32">
-  <img src="img/heart__full.svg" class="game__heart" alt="Life" width="32" height="32">
-</div>
-</header>
+import {footerData, gameState} from '../../data/data';
+import getHeader from './header';
+import getFooter from './footer';
+
+const template = String.raw`${getHeader(gameState)}
 <div class="game">
 <p class="game__task">Найдите рисунок среди изображений</p>
 <form class="game__content  game__content--triple">
@@ -45,16 +36,7 @@ const template = String.raw`<header class="header">
   </ul>
 </div>
 </div>
-<footer class="footer">
-<a href="https://htmlacademy.ru" class="social-link social-link--academy">HTML Academy</a>
-<span class="footer__made-in">Сделано в <a href="https://htmlacademy.ru" class="footer__link">HTML Academy</a> &copy; 2016</span>
-<div class="footer__social-links">
-  <a href="https://twitter.com/htmlacademy_ru" class="social-link  social-link--tw">Твиттер</a>
-  <a href="https://www.instagram.com/htmlacademy/" class="social-link  social-link--ins">Инстаграм</a>
-  <a href="https://www.facebook.com/htmlacademy" class="social-link  social-link--fb">Фэйсбук</a>
-  <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
-</div>
-</footer>`;
+${getFooter(footerData.creationDate)}`;
 
 const game3 = getElementFromTemplate(template);
 const linkToStartScreen = game3.querySelector(`.back`);
