+++ js/modules/game/rules.js
@@ -3,14 +3,11 @@ import renderScreen from '../utils/render-screen';
 import nextScreen from './game1';
 import intro from './intro';
 
-const template = String.raw`<header class="header">
-<div class="header__back">
-  <button class="back">
-    <img src="img/arrow_left.svg" width="45" height="45" alt="Back">
-    <img src="img/logo_small.svg" width="101" height="44">
-  </button>
-</div>
-</header>
+import {footerData} from '../../data/data';
+import getHeader from './header';
+import getFooter from './footer';
+
+const template = String.raw`${getHeader()}
 <div class="rules">
 <h1 class="rules__title">Правила</h1>
 <p class="rules__description">Угадай 10 раз для каждого изображения фото <img
@@ -27,16 +24,7 @@ const template = String.raw`<header class="header">
   <button class="rules__button  continue" type="submit" disabled>Go!</button>
 </form>
 </div>
-<footer class="footer">
-<a href="https://htmlacademy.ru" class="social-link social-link--academy">HTML Academy</a>
-<span class="footer__made-in">Сделано в <a href="https://htmlacademy.ru" class="footer__link">HTML Academy</a> &copy; 2016</span>
-<div class="footer__social-links">
-  <a href="https://twitter.com/htmlacademy_ru" class="social-link  social-link--tw">Твиттер</a>
-  <a href="https://www.instagram.com/htmlacademy/" class="social-link  social-link--ins">Инстаграм</a>
-  <a href="https://www.facebook.com/htmlacademy" class="social-link  social-link--fb">Фэйсбук</a>
-  <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
-</div>
-</footer>`;
+${getFooter(footerData.creationDate)}`;
 
 const rules = getElementFromTemplate(template);
 const rulesInput = rules.querySelector(`.rules__input`);
