+++ js/modules/game/stats.js
@@ -2,14 +2,11 @@ import getElementFromTemplate from '../utils/get-element-from-template';
 import renderScreen from '../utils/render-screen';
 import intro from './intro';
 
-const template = String.raw`<header class="header">
-<div class="header__back">
-  <button class="back">
-    <img src="img/arrow_left.svg" width="45" height="45" alt="Back">
-    <img src="img/logo_small.svg" width="101" height="44">
-  </button>
-</div>
-</header>
+import {footerData} from '../../data/data';
+import getHeader from './header';
+import getFooter from './footer';
+
+const template = String.raw`${getHeader()}
 <div class="result">
 <h1>Победа!</h1>
 <table class="result__table">
@@ -110,16 +107,7 @@ const template = String.raw`<header class="header">
   </tr>
 </table>
 </div>
-<footer class="footer">
-<a href="https://htmlacademy.ru" class="social-link social-link--academy">HTML Academy</a>
-<span class="footer__made-in">Сделано в <a href="https://htmlacademy.ru" class="footer__link">HTML Academy</a> &copy; 2016</span>
-<div class="footer__social-links">
-  <a href="https://twitter.com/htmlacademy_ru" class="social-link  social-link--tw">Твиттер</a>
-  <a href="https://www.instagram.com/htmlacademy/" class="social-link  social-link--ins">Инстаграм</a>
-  <a href="https://www.facebook.com/htmlacademy" class="social-link  social-link--fb">Фэйсбук</a>
-  <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
-</div>
-</footer>`;
+${getFooter(footerData.creationDate)}`;
 
 const stats = getElementFromTemplate(template);
 const linkToStartScreen = stats.querySelector(`.back`);
